+++ packages/client/src/components/Button/style.tsx
@@ -1,5 +1,6 @@
 import styled from 'styled-components';
 import { ButtonProps } from './Button';
+import { BOX_SHADOW_BUTTON } from 'styles/variables/styleConstants';
 
 export const StButton = styled.button`
   background: ${(props: ButtonProps) => {
@@ -30,8 +31,7 @@ export const StButton = styled.button`
     return props?.disignType === 'secondary' ? `1px solid #acb5bd` : `none`;
   }};
   border-radius: 1em;
-  box-shadow: 0 4px 4px rgba(0, 0, 0, 0.25),
-    inset 4px 4px 4px rgba(130, 128, 128, 0.25); //добавил по макету в фигме для экспереимента
+  box-shadow: ${BOX_SHADOW_BUTTON};
   font-weight: 700;
   text-align: center;
   font-size: ${(props: ButtonProps) => {
@@ -99,41 +99,4 @@ export const StButtonNewTopic = styled(StButton)`
   display: flex;
   align-items: center;
   justify-content: center;
-`;
-
-export const StButtonBackToProfile = styled(StButton)`
-  position: absolute;
-  left: 36px;
-  top: 50%;
-  transform: translateY(-50%);
-  z-index: 100;
-  background-color: initial;
-  border: none;
-  box-shadow: none;
-
-  &:hover svg {
-    filter: drop-shadow(0px 4px 4px rgba(0, 0, 0, 0.25));
-  }
-`;
-
-export const StButtonBackToForum = styled(StButton)`
-  position: absolute;
-  top: -42px;
-  left: 0;
-  z-index: 1;
-  font-weight: 500;
-  font-size: 20px;
-  line-height: 23px;
-  text-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
-  display: flex;
-  align-items: center;
-  text-align: center;
-  color: ${props => props.theme?.COLOR_TEXT_PRIMARY};
-  background-color: initial;
-  border: none;
-  box-shadow: none;
-
-  &:hover {
-    text-shadow: 0px 4px 4px rgb(0 0 0 / 40%);
-  }
 `;
